chore(Tests): split up e2e verifier tests
