+++ test/e2e/verifier/ethereum-main-invalid-merkle-root.test.js
@@ -1,0 +1,15 @@
+import { Certificate, SUB_STEPS, VERIFICATION_STATUSES } from '../../../src';
+import FIXTURES from '../../fixtures';
+
+describe('given the certificate is an ethereum main with an invalid merkle root', function () {
+  it('should fail', async function () {
+    const certificate = new Certificate(FIXTURES.EthereumMainInvalidMerkleRoot);
+    const result = await certificate.verify(({ code, label, status, errorMessage }) => {
+      if (code === SUB_STEPS.checkMerkleRoot && status !== VERIFICATION_STATUSES.STARTING) {
+        expect(status).toBe(VERIFICATION_STATUSES.FAILURE);
+        expect(errorMessage).toBe('Merkle root does not match remote hash.');
+      }
+    });
+    expect(result.status).toBe(VERIFICATION_STATUSES.FAILURE);
+  });
+});
