+++ test/e2e/verifier/etherum-ropsten-v2-valid.test.js
@@ -1,0 +1,10 @@
+import { Certificate, VERIFICATION_STATUSES } from '../../../src';
+import FIXTURES from '../../fixtures';
+
+describe('given the certificate is a valid ethereum ropsten', function () {
+  it('should verify successfully', async function () {
+    const certificate = new Certificate(FIXTURES.EthereumRopstenV2Valid);
+    const result = await certificate.verify();
+    expect(result.status).toBe(VERIFICATION_STATUSES.SUCCESS);
+  });
+});
