+++ test/e2e/verifier/mocknet-v2-valid.test.js
@@ -1,0 +1,10 @@
+import { Certificate, VERIFICATION_STATUSES } from '../../../src';
+import FIXTURES from '../../fixtures';
+
+describe('given the certificate is a valid mocknet (v2.0)', function () {
+  it('should verify successfully', async function () {
+    const certificate = new Certificate(FIXTURES.MocknetV2Valid);
+    const result = await certificate.verify();
+    expect(result.status).toBe(VERIFICATION_STATUSES.SUCCESS);
+  });
+});
